Integrate minimal Angular component for preview
Fixes #5
Add EditorPreview component that uses the JsonForms Angular webcomponent
for previewing the schema and uischema produced by the editor.

The Angular webcomponent provides a custom element '<ng-jsonforms>' that
uses JsonForms with the Angular Material renderer set to render a given
schema, UI schema and data (UI schema and data are optional).

The input data for the element can be set via the setJsonFormsInput
method, which we can access in the EditorPreview by using a ref.