Editable JSON and UI schemas
Integrate the Monaco Text Editor to make the JSON Schema and ui schema
editable.

The JSON schema and the new UI Schema tabs now feature a toolbar
offering copy and edit actions for their respective schemas. When edited
the Monaco editor is opened in a dialog. When the dialog is closed the
schemas are validated (at the moment only with a simple syntax check)
and set if successful.

Also:
 * Rename Schema panel to Palette
 * Integrates rescripts for CRA customizing (needed for Monaco editor
integration)