Editable properties for selected editor elements
Allows to select editor elements and edit their properties in the
Properties panel. The selected element is highlighted visually. The
properties are rendered with JSON Forms. The used schema and
ui schema are provided by a service based on the current selection.

To preserve selections in the case of ui schema changes UUIDs for
schema and ui schema elements are introduced. Also changes the
linking mechanism to use these UUIDs instead of paths.