Adjust JsonEditorDialog CSS
Control the dialog height via CSS on the root paper Material UI uses to
render its dialogs instead of using the DialogContent. This fixes an
issue where the dialog wouldn't be rendered with its intended height in
Firefox. To grow the MonacoEditor component an additional height
property must be set as react-monaco-editor uses a plain height approach
internally.