Include library sources in app build
Always use the library sources for the JSON Forms Editor app to
unify the development and production experience.

Also adds a 'testapp' package which simulates a regular Editor
consumer and showcases the minimal integration of the
'@jsonforms/editor' component. The testapp is excluded from the
regular build but is automatically smoketested by the CI process.