Allow customizing properties and editor view, fix Monaco editors (#111)
- allow clients to provide their own renderers for the properties view
- allow clients to provide their own renderers for the editor view
- allow customizing the icon of EditorElements. This is useful e.g. for
client specific schema elements.
- fix Monaco editor integration. We use the Monaco editor in different
places in the UI (JSON Schema view, UI Schema view, Rule editor in the
Properties View). All these instances of the editor are using different
models for the 'json' language, but we were modifying the same model
which is shared between instances. As a result, the schema of the last
opened editor was used in all instances of the Monaco editor. We fixed
this by using the fileMatch filter to match the correct schema for the
editor instance, based on the unique uri of each model.